routes updated and pages renamed accordingly
