+++ components/public/nav.js
@@ -7,19 +7,19 @@ const Nav = ({ page, projectTitle }) => (
   <nav>
     {page === 'home' ? (
       <>
-        <Link href="/portfolio">
+        <Link href="/=">
           <a className="nav-link--first" style={{ color: '#fff' }}>
             Work
           </a>
         </Link>
-        <Link href="/about/">
+        <Link href="/about">
           <a style={{ color: '#fff', opacity: 0.4 }}>Info</a>
         </Link>
         <div className="project-title">{projectTitle}</div>
       </>
     ) : (
       <>
-        <Link href="/portfolio">
+        <Link href="/=">
           <a
             className="nav-link--first"
             style={{ color: '#0d0d0d', opacity: 0.4 }}
@@ -27,7 +27,7 @@ const Nav = ({ page, projectTitle }) => (
             Work
           </a>
         </Link>
-        <Link href="/about/">
+        <Link href="/about">
           <a style={{ color: '#0d0d0d' }}>Info</a>
         </Link>
         <Contact />
