+++ package.json
@@ -22,6 +22,7 @@
     "dotenv": "^7.0.0",
     "express": "^4.16.4",
     "express-session": "^1.16.2",
+    "flickity-bg-lazyload": "^1.0.1",
     "isomorphic-fetch": "^2.2.1",
     "mongodb": "^3.2.3",
     "next": "^8.1.0",
